+++ Cargo.toml
@@ -115,6 +115,3 @@ inherits = "test"
 debug = 0 # for saving disk space during linking
 incremental = false
 codegen-units = 16
-
-#[patch."https://github.com/brson/rbpf.git"]
-#solana_rbpf.path = "../rbpf"
