+++ language/tools/move-mv-llvm-compiler/llvm-extra-sys/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
 name = "llvm-extra-sys"
 version = "0.1.0"
+authors = ["Diem Association <opensource@diem.com>"]
+description = "Additional LLVM glue"
 license = "Apache 2.0 / MIT"
+publish = false
 edition = "2021"
 
 [dependencies]
