+++ language/tools/move-mv-llvm-compiler/tests/move-ir-tests/assign-build/scripts/main.expected.ll
@@ -2,8 +2,8 @@
 source_filename = "<unknown>"
 
 define void @main() {
-entry_:
-  %local_0_ = alloca i8, align 1
-  store i8 7, ptr %local_0_, align 1
+entry:
+  %local_0 = alloca i8, align 1
+  store i8 7, ptr %local_0, align 1
   ret void
 }
