+++ language/tools/move-mv-llvm-compiler/tests/move-ir-tests/if-build/modules/0_Test.expected.ll
@@ -2,34 +2,34 @@
 source_filename = "<unknown>"
 
 define i8 @test(i1 %0) {
-entry_:
-  %local_0_ = alloca i1, align 1
-  %local_1_ = alloca i8, align 1
-  %local_2_ = alloca i1, align 1
-  %local_3_ = alloca i8, align 1
-  %local_4_ = alloca i8, align 1
-  %local_5_ = alloca i8, align 1
-  store i1 %0, ptr %local_0_, align 1
-  %load_store_tmp_ = load i1, ptr %local_0_, align 1
-  store i1 %load_store_tmp_, ptr %local_2_, align 1
-  %cnd_ = load i1, ptr %local_2_, align 1
-  br i1 %cnd_, label %bb_1_, label %bb_0_
+entry:
+  %local_0 = alloca i1, align 1
+  %local_1 = alloca i8, align 1
+  %local_2 = alloca i1, align 1
+  %local_3 = alloca i8, align 1
+  %local_4 = alloca i8, align 1
+  %local_5 = alloca i8, align 1
+  store i1 %0, ptr %local_0, align 1
+  %load_store_tmp = load i1, ptr %local_0, align 1
+  store i1 %load_store_tmp, ptr %local_2, align 1
+  %cnd = load i1, ptr %local_2, align 1
+  br i1 %cnd, label %bb_1, label %bb_0
 
-bb_1_:                                            ; preds = %entry_
-  store i8 2, ptr %local_3_, align 1
-  %load_store_tmp_1 = load i8, ptr %local_3_, align 1
-  store i8 %load_store_tmp_1, ptr %local_1_, align 1
-  br label %bb_2_
+bb_1:                                             ; preds = %entry
+  store i8 2, ptr %local_3, align 1
+  %load_store_tmp1 = load i8, ptr %local_3, align 1
+  store i8 %load_store_tmp1, ptr %local_1, align 1
+  br label %bb_2
 
-bb_0_:                                            ; preds = %entry_
-  store i8 3, ptr %local_4_, align 1
-  %load_store_tmp_2 = load i8, ptr %local_4_, align 1
-  store i8 %load_store_tmp_2, ptr %local_1_, align 1
-  br label %bb_2_
+bb_0:                                             ; preds = %entry
+  store i8 3, ptr %local_4, align 1
+  %load_store_tmp2 = load i8, ptr %local_4, align 1
+  store i8 %load_store_tmp2, ptr %local_1, align 1
+  br label %bb_2
 
-bb_2_:                                            ; preds = %bb_0_, %bb_1_
-  %load_store_tmp_3 = load i8, ptr %local_1_, align 1
-  store i8 %load_store_tmp_3, ptr %local_5_, align 1
-  %retval_ = load i8, ptr %local_5_, align 1
-  ret i8 %retval_
+bb_2:                                             ; preds = %bb_0, %bb_1
+  %load_store_tmp3 = load i8, ptr %local_1, align 1
+  store i8 %load_store_tmp3, ptr %local_5, align 1
+  %retval = load i8, ptr %local_5, align 1
+  ret i8 %retval
 }
