+++ language/tools/move-mv-llvm-compiler/tests/move-ir-tests/return-build/modules/0_Test.expected.ll
@@ -2,9 +2,9 @@
 source_filename = "<unknown>"
 
 define i8 @test() {
-entry_:
-  %local_0_ = alloca i8, align 1
-  store i8 100, ptr %local_0_, align 1
-  %retval_ = load i8, ptr %local_0_, align 1
-  ret i8 %retval_
+entry:
+  %local_0 = alloca i8, align 1
+  store i8 100, ptr %local_0, align 1
+  %retval = load i8, ptr %local_0, align 1
+  ret i8 %retval
 }
