+++ language/tools/move-mv-llvm-compiler/tests/move-ir-tests/script-build/scripts/main.expected.ll
@@ -2,6 +2,6 @@
 source_filename = "<unknown>"
 
 define void @main() {
-entry_:
+entry:
   ret void
 }
