+++ language/tools/move-mv-llvm-compiler/tests/move-ir-tests/script.move
@@ -1,4 +1,3 @@
-// ignore
 script {
   fun main() { }
 }
